move schemas to separate file
